Rename `AbstractBackend` to `AbstractJulBackend`, and move it to `backend.jul` package
