+++ backends/logging-jul-backend/src/main/kotlin/io/spine/logging/backend/jul/AbstractJulBackend.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014, The Flogger Authors; 2023, TeamDev. All rights reserved.
+ * Copyright 2023, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.backend.system;
+package io.spine.logging.backend.jul;
 
 import io.spine.logging.flogger.backend.LoggerBackend;
 
@@ -35,13 +35,15 @@ import java.util.logging.LogRecord;
 import java.util.logging.Logger;
 
 /**
- * Common backend to handle everything except formatting of log message and metadata. This is an
- * unstable implementation and should not be used outside of the Flogger core library.
+ * Common backend to handle everything except formatting of a log message
+ * and metadata.
+ *
+ * This is an unstable implementation and should not be used outside of the Flogger core library.
  *
  * @see <a href="https://github.com/google/flogger/blob/cb9e836a897d36a78309ee8badf5cad4e6a2d3d8/api/src/main/java/com/google/common/flogger/backend/system/AbstractBackend.java">
  *     Original Java code of Google Flogger</a>
  */
-public abstract class AbstractBackend extends LoggerBackend {
+public abstract class AbstractJulBackend extends LoggerBackend {
   // Set if any attempt at logging via the "forcing" logger fails due to an inability to set the
   // log level in the forcing logger. This result is cached so we don't repeatedly trigger
   // security exceptions every time something is logged. This field is only ever read or written
@@ -54,7 +56,7 @@ public abstract class AbstractBackend extends LoggerBackend {
   // class name. This needs work to handle anonymous loggers however (if that's ever supported).
   // TODO:2023-09-14:yevhenii.nadtochii: Should become `internal` when migrated to Kotlin.
   // See issue: https://github.com/SpineEventEngine/logging/issues/47
-  protected AbstractBackend(Logger logger) {
+  protected AbstractJulBackend(Logger logger) {
     this.logger = logger;
   }
 
@@ -67,7 +69,7 @@ public abstract class AbstractBackend extends LoggerBackend {
    * the outer class name. There is no benefit to having loggers named after an inner/nested
    * classes, and this distinction is expected to go away.
    */
-  protected AbstractBackend(String loggingClass) {
+  protected AbstractJulBackend(String loggingClass) {
     // TODO(b/27920233): Strip inner/nested classes when deriving logger name.
     this(Logger.getLogger(loggingClass.replace('$', '.')));
   }
